finished services, need to test and finish menu

+++ src/entity/Author.java
@@ -1,20 +1,18 @@
 package entity;
 
 public class Author {
-	private long authorID;
+	private int authorID;
 	private String name;
 	//private String lastName;
-	public Author(long authorID, String name) {
+	public Author(int authorID, String name) {
 		super();
 		this.authorID = authorID;
 		this.name = name;
 	}
-	
-	public long getAuthorID() {
+	public int getAuthorID() {
 		return authorID;
 	}
-
-	public void setAuthorID(long authorID) {
+	public void setAuthorID(int authorID) {
 		this.authorID = authorID;
 	}
 	public String getName() {
@@ -23,6 +21,7 @@ public class Author {
 	public void setName(String name) {
 		this.name = name;
 	}
+	
 
 	
 }
